Relocate to Chars and share
